moved persistence to internal
